move getUseCaseFeatureConfig and isNavGroupInFeatureConfigs to core
