+++ src/core/utils/constants.ts
@@ -6,3 +6,4 @@
 export const WORKSPACE_TYPE = 'workspace';
 
 export const WORKSPACE_PATH_PREFIX = '/w';
+export const WORKSPACE_USE_CASE_PREFIX = 'use-case-';
