+++ src/core/utils/index.ts
@@ -37,8 +37,14 @@ export {
   IContextProvider,
 } from './context';
 export { DEFAULT_APP_CATEGORIES } from './default_app_categories';
-export { WORKSPACE_PATH_PREFIX, WORKSPACE_TYPE } from './constants';
-export { getWorkspaceIdFromUrl, formatUrlWithWorkspaceId, cleanWorkspaceId } from './workspace';
+export { WORKSPACE_USE_CASE_PREFIX, WORKSPACE_PATH_PREFIX, WORKSPACE_TYPE } from './constants';
+export {
+  getWorkspaceIdFromUrl,
+  formatUrlWithWorkspaceId,
+  cleanWorkspaceId,
+  getUseCaseFeatureConfig,
+  isNavGroupInFeatureConfigs,
+} from './workspace';
 export {
   DEFAULT_NAV_GROUPS,
   ALL_USE_CASE_ID,
