+++ src/plugins/workspace/common/constants.ts
@@ -8,6 +8,7 @@ import {
   DATA_CONNECTION_SAVED_OBJECT_TYPE,
   DATA_SOURCE_SAVED_OBJECT_TYPE,
 } from '../../data_source/common';
+
 export const WORKSPACE_FATAL_ERROR_APP_ID = 'workspace_fatal_error';
 export const WORKSPACE_CREATE_APP_ID = 'workspace_create';
 export const WORKSPACE_LIST_APP_ID = 'workspace_list';
@@ -162,7 +163,6 @@ export enum AssociationDataSourceModalMode {
   OpenSearchConnections = 'opensearch-connections',
   DirectQueryConnections = 'direction-query-connections',
 }
-export const USE_CASE_PREFIX = 'use-case-';
 export const OPENSEARCHDASHBOARDS_CONFIG_PATH = 'opensearchDashboards';
 
 // Workspace will handle both data source and data connection type saved object.
