+++ src/plugins/workspace/public/index.ts
@@ -11,4 +11,3 @@ export function plugin() {
 
 export { WorkspacePluginSetup, WorkspaceCollaborator } from './types';
 export { WorkspaceCollaboratorType } from './services';
-export { isNavGroupInFeatureConfigs } from './utils';
